changed type of max_capacity

+++ BackEnd/criminal/src/main/Java/com/desarrollo/criminal/dto/request/UpdatePATCHAppointmentDTO.java
@@ -20,7 +20,7 @@ public class UpdatePATCHAppointmentDTO {
 
     private Long activityID;
 
-    private Long max_capacity;
+    private Integer max_capacity;
 
     private Long instructorID;
